requirements.txt
# Production dependencies extracted from pyproject.toml
Django>=5.0.0
psycopg2-binary>=2.9.9
redis>=5.0.1
celery[redis]>=5.3.4
django-celery-beat>=2.5.0
django-celery-results>=2.5.1
django-allauth>=0.57.0
django-filter>=23.5
django-tables2>=2.7.0
django-braces>=1.15.0
django-flags>=5.0.12
django-extensions>=3.2.3
django-environ>=0.11.2
django-cors-headers>=4.3.1
django-csp>=3.7
django-guardian>=2.4.0
djangorestframework>=3.14.0
drf-spectacular>=0.27.0
stripe>=7.8.0
cryptography>=41.0.7
pydantic>=2.5.2
requests>=2.31.0
python-dotenv>=1.0.0
whitenoise>=6.6.0
gunicorn>=21.2.0
sentry-sdk[django]>=1.39.1
prometheus-client>=0.19.0
django-prometheus>=2.3.1
boto3>=1.34.0
scikit-learn>=1.3.2
xgboost>=2.0.2
lightgbm>=4.1.0
prophet>=1.1.5
matplotlib>=3.8.2
plotly>=5.17.0
kaleido>=0.2.1
pandas>=2.1.4
numpy>=1.26.2
scipy>=1.11.4
dj-database-url>=2.1.0
django-storages>=1.14.2
pillow>=10.1.0
python-dateutil>=2.8.2
pytz>=2023.3
babel>=2.14.0
django-timezone-field>=6.1.0
PyJWT>=2.8.0

# Development dependencies
django-debug-toolbar>=4.2.0
python-json-logger
